+++ day17/bitwise_xor_operation_test.go
@@ -6,21 +6,21 @@ import (
 )
 
 func TestTwoZeros(t *testing.T) {
-	assert.Equal(t, int64(0), BitwiseXor(0, 0))
+	assert.Equal(t, 0, BitwiseXor(0, 0))
 }
 
 func TestZeroAndOne(t *testing.T) {
-	assert.Equal(t, int64(1), BitwiseXor(0, 1))
-	assert.Equal(t, int64(1), BitwiseXor(1, 0))
+	assert.Equal(t, 1, BitwiseXor(0, 1))
+	assert.Equal(t, 1, BitwiseXor(1, 0))
 }
 
 func TestTwoOne(t *testing.T) {
-	assert.Equal(t, int64(0), BitwiseXor(1, 1))
+	assert.Equal(t, 0, BitwiseXor(1, 1))
 }
 
 func TestProvidedExamples(t *testing.T) {
-	assert.Equal(t, int64(26), BitwiseXor(29, 7))
-	assert.Equal(t, int64(26), BitwiseXor(7, 29))
-	assert.Equal(t, int64(44354), BitwiseXor(2024, 43690))
-	assert.Equal(t, int64(44354), BitwiseXor(43690, 2024))
+	assert.Equal(t, 26, BitwiseXor(29, 7))
+	assert.Equal(t, 26, BitwiseXor(7, 29))
+	assert.Equal(t, 44354, BitwiseXor(2024, 43690))
+	assert.Equal(t, 44354, BitwiseXor(43690, 2024))
 }
